feat(auth): extract auto-login and Google login to custom hooks
Refactored LoginPage to use new useAutoLogin and useGoogleLogin hooks.
This improves separation of concerns and code readability by moving
side-effect logic out of the component and into reusable hooks. No
functional changes to authentication flow.
